+++ packages/reactive_ble_mobile/android/src/main/kotlin/com/signify/hue/flutterreactiveble/ble/BleClient.kt
@@ -75,5 +75,5 @@ interface BleClient {
         value: ByteArray
     )
 
-    fun isDeviceConnected(deviceId: String): Observable<EstablishConnectionResult>
+    fun isDeviceConnected(deviceId: String): Boolean
 }